chore(j-s): All types of investigation cases can now have multiple defendants (#7818)
* Refactor defendant/defendantForm into defendant

* Add the ability to have multiple defendants on all case types in investigation cases

* Resolve naming conflict

* Rename D to TDefendant
