+++ apps/judicial-system/web/src/routes/Prosecutor/InvestigationCase/Defendant/Defendant.tsx
@@ -1,25 +1,47 @@
 import React, { useContext } from 'react'
 import { useRouter } from 'next/router'
 import { useIntl } from 'react-intl'
+import { ValueType } from 'react-select/src/types'
+import { AnimatePresence, motion } from 'framer-motion'
+import { uuid } from 'uuidv4'
 
-import { PageLayout } from '@island.is/judicial-system-web/src/components'
+import {
+  BlueBox,
+  FormContentContainer,
+  FormFooter,
+  PageLayout,
+} from '@island.is/judicial-system-web/src/components'
 import useDefendants from '@island.is/judicial-system-web/src/utils/hooks/useDefendants'
 import {
   ProsecutorSubsections,
+  ReactSelectOption,
   Sections,
 } from '@island.is/judicial-system-web/src/types'
 import { useCase } from '@island.is/judicial-system-web/src/utils/hooks'
 import { FormContext } from '@island.is/judicial-system-web/src/components/FormProvider/FormProvider'
 import PageHeader from '@island.is/judicial-system-web/src/components/PageHeader/PageHeader'
-import { titles } from '@island.is/judicial-system-web/messages'
-import type { Case } from '@island.is/judicial-system/types'
+import { titles, defendant as m } from '@island.is/judicial-system-web/messages'
+import {
+  Case,
+  CaseType,
+  Defendant as TDefendant,
+  UpdateDefendant,
+} from '@island.is/judicial-system/types'
+import { Box, Button, Input, Select, Text } from '@island.is/island-ui/core'
+import { isDefendantStepValidIC } from '@island.is/judicial-system-web/src/utils/validate'
+import DefenderInfo from '@island.is/judicial-system-web/src/components/DefenderInfo/DefenderInfo'
+import { setAndSendToServer } from '@island.is/judicial-system-web/src/utils/formHelper'
+import { capitalize, caseTypes } from '@island.is/judicial-system/formatters'
+import { theme } from '@island.is/island-ui/theme'
 import * as constants from '@island.is/judicial-system/consts'
 
-import DefendantForm from './DefendantForm'
+import LokeCaseNumber from '../../SharedComponents/LokeCaseNumber/LokeCaseNumber'
+import DefendantInfo from '../../SharedComponents/DefendantInfo/DefendantInfo'
+import { isBusiness } from '@island.is/judicial-system-web/src/utils/stepHelper'
 
 const Defendant = () => {
   const router = useRouter()
-  const { updateDefendant, createDefendant } = useDefendants()
+  const { updateDefendant, createDefendant, deleteDefendant } = useDefendants()
 
   const {
     workingCase,
@@ -27,7 +49,7 @@ const Defendant = () => {
     isLoadingWorkingCase,
     caseNotFound,
   } = useContext(FormContext)
-  const { createCase, isCreatingCase } = useCase()
+  const { createCase, isCreatingCase, updateCase } = useCase()
   const { formatMessage } = useIntl()
 
   const handleNextButtonClick = async (theCase: Case) => {
@@ -76,6 +98,104 @@ const Defendant = () => {
     }
   }
 
+  const updateDefendantState = (
+    defendantId: string,
+    update: UpdateDefendant,
+  ) => {
+    if (workingCase.defendants) {
+      const indexOfDefendantToUpdate = workingCase.defendants.findIndex(
+        (defendant) => defendant.id === defendantId,
+      )
+
+      const newDefendants = [...workingCase.defendants]
+
+      newDefendants[indexOfDefendantToUpdate] = {
+        ...newDefendants[indexOfDefendantToUpdate],
+        ...update,
+      }
+
+      setWorkingCase({ ...workingCase, defendants: newDefendants })
+    }
+  }
+
+  const handleUpdateDefendant = async (
+    defendantId: string,
+    updatedDefendant: UpdateDefendant,
+  ) => {
+    updateDefendantState(defendantId, updatedDefendant)
+
+    if (workingCase.id) {
+      updateDefendant(workingCase.id, defendantId, updatedDefendant)
+    }
+  }
+
+  const handleDeleteDefendant = async (defendant: TDefendant) => {
+    if (workingCase.defendants && workingCase.defendants.length > 1) {
+      if (workingCase.id) {
+        const defendantDeleted = await deleteDefendant(
+          workingCase.id,
+          defendant.id,
+        )
+
+        if (defendantDeleted && workingCase.defendants) {
+          removeDefendantFromState(defendant)
+        } else {
+          // TODO: handle error
+        }
+      } else {
+        removeDefendantFromState(defendant)
+      }
+    }
+  }
+
+  const removeDefendantFromState = (defendant: TDefendant) => {
+    if (workingCase.defendants && workingCase.defendants?.length > 1) {
+      setWorkingCase({
+        ...workingCase,
+        defendants: [...workingCase.defendants].filter(
+          (d) => d.id !== defendant.id,
+        ),
+      })
+    }
+  }
+
+  const handleCreateDefendantClick = async () => {
+    if (workingCase.id) {
+      const defendantId = await createDefendant(workingCase.id, {
+        gender: undefined,
+        name: '',
+        address: '',
+        nationalId: '',
+        citizenship: '',
+      })
+
+      createEmptyDefendant(defendantId)
+    } else {
+      createEmptyDefendant()
+    }
+
+    window.scrollTo(0, document.body.scrollHeight)
+  }
+
+  const createEmptyDefendant = (defendantId?: string) => {
+    if (workingCase.defendants) {
+      setWorkingCase({
+        ...workingCase,
+        defendants: [
+          ...workingCase.defendants,
+          {
+            id: defendantId || uuid(),
+            gender: undefined,
+            name: '',
+            nationalId: '',
+            address: '',
+            citizenship: '',
+          } as TDefendant,
+        ],
+      })
+    }
+  }
+
   return (
     <PageLayout
       workingCase={workingCase}
@@ -90,12 +210,173 @@ const Defendant = () => {
       <PageHeader
         title={formatMessage(titles.prosecutor.investigationCases.defendant)}
       />
-      <DefendantForm
-        workingCase={workingCase}
-        setWorkingCase={setWorkingCase}
-        handleNextButtonClick={handleNextButtonClick}
-        isLoading={isCreatingCase}
-      />
+
+      <FormContentContainer>
+        <Box marginBottom={10}>
+          <Box marginBottom={7}>
+            <Text as="h1" variant="h1">
+              {formatMessage(m.heading)}
+            </Text>
+          </Box>
+          <Box component="section" marginBottom={5}>
+            <LokeCaseNumber
+              workingCase={workingCase}
+              setWorkingCase={setWorkingCase}
+            />
+          </Box>
+          <Box component="section" marginBottom={5}>
+            <Box marginBottom={3}>
+              <Text as="h3" variant="h3">
+                {formatMessage(m.sections.investigationType.heading)}
+              </Text>
+            </Box>
+            <BlueBox>
+              <Box marginBottom={3}>
+                <Select
+                  name="type"
+                  options={constants.ICaseTypes as ReactSelectOption[]}
+                  label={formatMessage(m.sections.investigationType.type.label)}
+                  placeholder={formatMessage(
+                    m.sections.investigationType.type.placeholder,
+                  )}
+                  onChange={(selectedOption: ValueType<ReactSelectOption>) =>
+                    setAndSendToServer(
+                      'type',
+                      (selectedOption as ReactSelectOption).value as string,
+                      workingCase,
+                      setWorkingCase,
+                      updateCase,
+                    )
+                  }
+                  value={
+                    workingCase?.id
+                      ? {
+                          value: CaseType[workingCase.type],
+                          label: capitalize(caseTypes[workingCase.type]),
+                        }
+                      : undefined
+                  }
+                  formatGroupLabel={() => (
+                    <div
+                      style={{
+                        width: 'calc(100% + 24px)',
+                        height: '3px',
+                        marginLeft: '-12px',
+                        backgroundColor: theme.color.dark300,
+                      }}
+                    />
+                  )}
+                  required
+                />
+              </Box>
+              <Input
+                data-testid="description"
+                name="description"
+                label={formatMessage(
+                  m.sections.investigationType.description.label,
+                )}
+                placeholder={formatMessage(
+                  m.sections.investigationType.description.placeholder,
+                )}
+                value={workingCase.description || ''}
+                autoComplete="off"
+                onChange={(evt) => {
+                  setWorkingCase({
+                    ...workingCase,
+                    description: evt.target.value,
+                  })
+                }}
+                onBlur={(evt) =>
+                  setAndSendToServer(
+                    'description',
+                    evt.target.value,
+                    workingCase,
+                    setWorkingCase,
+                    updateCase,
+                  )
+                }
+              />
+            </BlueBox>
+          </Box>
+          <Box component="section" marginBottom={5}>
+            <Box marginBottom={3}>
+              <Text as="h3" variant="h3">
+                {formatMessage(m.sections.defendantInfo.heading)}
+              </Text>
+            </Box>
+            <AnimatePresence>
+              {workingCase.defendants &&
+                workingCase.defendants.map((defendant, index) => (
+                  <motion.div
+                    key={index}
+                    initial={{ opacity: 0, y: 10 }}
+                    animate={{ opacity: 1, y: 0 }}
+                    exit={{ opacity: 0, y: 10 }}
+                  >
+                    <Box
+                      marginBottom={
+                        index - 1 === workingCase.defendants?.length ? 0 : 3
+                      }
+                    >
+                      <DefendantInfo
+                        defendant={defendant}
+                        onDelete={index > 0 ? handleDeleteDefendant : undefined}
+                        onChange={handleUpdateDefendant}
+                        updateDefendantState={updateDefendantState}
+                      />
+                    </Box>
+                  </motion.div>
+                ))}
+            </AnimatePresence>
+            <Box display="flex" justifyContent="flexEnd" marginTop={3}>
+              <Button
+                variant="ghost"
+                icon="add"
+                onClick={handleCreateDefendantClick}
+                disabled={workingCase.defendants?.some(
+                  (defendant) =>
+                    (!isBusiness(defendant.nationalId) && !defendant.gender) ||
+                    !defendant.name ||
+                    !defendant.address ||
+                    (!defendant.noNationalId && !defendant.nationalId),
+                )}
+              >
+                {formatMessage(m.sections.defendantInfo.addDefendantButtonText)}
+              </Button>
+            </Box>
+          </Box>
+          <AnimatePresence>
+            {[
+              CaseType.RESTRAINING_ORDER,
+              CaseType.EXPULSION_FROM_HOME,
+              CaseType.PSYCHIATRIC_EXAMINATION,
+              CaseType.OTHER,
+            ].includes(workingCase.type) && (
+              <motion.section
+                initial={{ opacity: 0, y: 10 }}
+                animate={{ opacity: 1, y: 0 }}
+                exit={{ opacity: 0, y: 10 }}
+              >
+                <DefenderInfo
+                  workingCase={workingCase}
+                  setWorkingCase={setWorkingCase}
+                />
+              </motion.section>
+            )}
+          </AnimatePresence>
+        </Box>
+      </FormContentContainer>
+      <FormContentContainer isFooter>
+        <FormFooter
+          previousUrl={`${constants.CASE_LIST_ROUTE}`}
+          onNextButtonClick={() => handleNextButtonClick(workingCase)}
+          nextIsDisabled={!isDefendantStepValidIC(workingCase)}
+          nextIsLoading={isCreatingCase}
+          nextButtonText={
+            workingCase.id === '' ? 'Stofna kröfu' : 'Halda áfram'
+          }
+        />
+      </FormContentContainer>
     </PageLayout>
   )
 }
