+++ libs/judicial-system/types/src/index.ts
@@ -53,7 +53,6 @@ export {
   isAcceptingCaseDecision,
   completedCaseStates,
   hasCaseBeenAppealed,
-  isCaseTypeWithMultipleDefendantsSupport,
 } from './lib/case'
 export type {
   Case,
